+++ src/Text/Hakyll/CreateContext.hs
@@ -5,37 +5,40 @@ module Text.Hakyll.CreateContext
     ( createPage
     , createCustomPage
     , createListing
+    , addField
     , combine
     , combineWithUrl
     ) where
 
+import Prelude hiding (id)
+
 import qualified Data.Map as M
-import Control.Arrow (second)
+import Control.Arrow (second, arr, (&&&), (***))
 import Control.Monad (liftM2)
 import Control.Applicative ((<$>))
+import Control.Arrow ((>>>))
+import Control.Category (id)
 
-import Text.Hakyll.File
 import Text.Hakyll.Context
 import Text.Hakyll.HakyllAction
 import Text.Hakyll.Render
-import Text.Hakyll.Internal.Page
+import Text.Hakyll.Page
+import Text.Hakyll.Pandoc
+import Text.Hakyll.Internal.Cache
 
 -- | Create a @Context@ from a page file stored on the disk. This is probably
 --   the most common way to create a @Context@.
 createPage :: FilePath -> HakyllAction () Context
-createPage path = HakyllAction
-    { actionDependencies = [path]
-    , actionUrl          = Left $ toUrl path
-    , actionFunction     = const (readPage path)
-    }
+createPage path = cacheAction "pages" $ readPageAction path >>> renderAction
 
--- | Create a "custom page" @Context@.
+-- | Create a custom page @Context@.
 --   
 --   The association list given maps keys to values for substitution. Note
 --   that as value, you can either give a @String@ or a
 --   @HakyllAction () String@. The latter is preferred for more complex data,
 --   since it allows dependency checking. A @String@ is obviously more simple
 --   to use in some cases.
+--
 createCustomPage :: FilePath
                  -> [(String, Either String (HakyllAction () String))]
                  -> HakyllAction () Context
@@ -68,6 +71,18 @@ createListing url templates renderables additional =
   where
     context = ("body", Right concatenation) : additional
     concatenation = renderAndConcat templates renderables
+
+-- | Add a field to a 'Context'.
+--
+addField :: String                                  -- ^ Key
+         -> Either String (HakyllAction () String)  -- ^ Value
+         -> HakyllAction Context Context            -- ^ Result
+addField key value =   arr (const ()) &&& id
+                   >>> value' *** id
+                   >>> arr (uncurry insert)
+  where
+    value' = arr (const ()) >>> either (arr . const) id value
+    insert v = Context . M.insert key v . unContext
 
 -- | Combine two @Context@s. The url will always be taken from the first
 --   @Renderable@. Also, if a `$key` is present in both renderables, the